+++ src/insert/insert_state.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, PartialEq)]
+pub enum InsertState {
+	Prompt,
+	Edit,
+}